requirements.txt
Django
sentence-transformers
faiss-cpu
numpy
groq
python-dotenv
